+++ kaithem/src/astrallibwrapper.py
@@ -16,42 +16,45 @@
 """This file is an wrapper around some of the astral library, a pure python
 library providing all of kaithem's astronomical functions"""
 
-import astral, pytz,datetime,time,calendar
-a = astral.Astral()
+import pytz,datetime,time,calendar
+from astral import LocationInfo
+import astral.sun
+import astral.moon
 
-def dawn(lat,lon,date=None):
+def getLocationInfo(lat,lon):
+    return LocationInfo("unknown","unknown","Etc/UTC", lat,lon)
+
+def dawn(lat,lon,date=None,elevation=0):
     "Given a latitude and longitude, return civil dawn time for any given date object as a unix timestamp(default to today)"
     if date==None:
         date= datetime.datetime.utcnow().date()
 
-    return calendar.timegm(a.dawn_utc(date,lat,lon).timetuple())
+    return calendar.timegm(astral.sun.sun(astral.Observer(lat,lon,elevation), date)['dawn'].timetuple())
 
-def dusk(lat,lon,date=None):
+def dusk(lat,lon,date=None,elevation=0):
     "Given a latitude and longitude, return civil dusk time for any given date object as a unix timestamp(default to today)"
     if date==None:
         date= datetime.datetime.utcnow().date()
 
-    return calendar.timegm(a.dusk_utc(date,lat,lon).timetuple())
+    return calendar.timegm(astral.sun.sun(astral.Observer(lat,lon,elevation), date)['dusk'].timetuple())
 
-def sunrise(lat,lon,date=None):
+def sunrise(lat,lon,date=None,elevation=0):
     "Given a latitude and longitude, return sunrise time for any given date object as a unix timestamp(default to today)"
     if date==None:
         date= datetime.datetime.utcnow().date()
 
-    return calendar.timegm(a.sunrise_utc(date,lat,lon).timetuple())
+    return calendar.timegm(astral.sun.sun(astral.Observer(lat,lon,elevation), date)['sunrise'].timetuple())
 
-def sunset(lat,lon,date=None):
-    "Given a latitude and longitude, return sunrise time for any given date object as a unix timestamp(default to today)"
+def sunset(lat,lon,date=None,elevation=0):
     if date==None:
         date= datetime.datetime.utcnow().date()
+    return calendar.timegm(astral.sun.sun(astral.Observer(lat,lon,elevation), date)['sunset'].timetuple())
 
-    return calendar.timegm(a.sunset_utc(date,lat,lon).timetuple())
-
-def rahu(lat,lon,date=None):
+def rahu(lat,lon,date=None,elevation=0):
     "Given a latitude and longitude, return a tuple of the start and end timestamps of the given date's rahukalaam period"
     if date==None:
         date= datetime.datetime.utcnow().date()
-    r = a.rahukaalam_utc(date,lat,lon)
+    r = astral.rahukaalam(astral.Observer(lat,lon,elevation),date)
     return (calendar.timegm(r['start'].timetuple()) , calendar.timegm(r['end'].timetuple()))
 
 
@@ -68,17 +71,18 @@ def isLight(lat,lon):
     return (dawn(lat,lon) <= time.time() <= dusk(lat,lon))
 
 def isRahu(lat,lon):
+    
         return (rahu(lat,lon)[0] <= time.time() <= rahu(lat,lon)[1])
 
 def moon():
     """
-        return 0 to 21 depending on current moon phase..
+        return 0 to 28 depending on current moon phase..
                 | 0  = New moon
                 | 7  = First quarter
                 | 14 = Full moon
                 | 21 = Last quarter
    """
-    return a.moon_phase(datetime.date.today(),pytz.UTC)
+    return astral.moon.phase(datetime.datetime.utcnow())
 
 seasons= {"spring": 0, "summer":1, "fall":2,"autumn":2, "winter": 3, 1:1, 2:2, 3:3, 4:4, 5:5}
 seasonnames = ["spring","summer","autumn","winter"]